+++ AvgScript/src/extensions/arrayEx.ts
@@ -7,6 +7,10 @@ declare global {
         empty(): boolean;
         hasValue(item: T): boolean;
         findIf(cmp: (item: T) => boolean): number;
+
+        removeIf(cmp: (item: T) => boolean): boolean;
+        remove(elementToRemove: T): boolean;
+
         uniquePush(elementToPush: T): number;
         uniquePushIf(elementToPush: T, cmp: (l: T, r: T) => boolean): number;
     }
@@ -39,6 +43,26 @@ Array.prototype.findIf = function <T>(cmp: (item: T) => boolean) {
     }
 
     return elementPosition;
+};
+
+Array.prototype.removeIf = function <T>(cmp: (item: T) => boolean) {
+    let pos = this.findIf(cmp);
+
+    if (pos === -1) {
+        return false;
+    }
+
+    this[pos] = this[this.length - 1];
+    this.pop();
+
+    return true;
+};
+
+
+Array.prototype.remove = function <T>(elementToRemove: T) {
+    return this.removeIf((item: T) => {
+        return comparator(item, elementToRemove);
+    });
 };
 
 // always return the element position in array